requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pypdf==4.3.1
Pillow==10.4.0
pytesseract==0.3.13
pytest==8.3.2
openai==1.106.1
httpx==0.28.1
langchain-openai==0.3.33
